pyproject.toml
[project]
name = "cnn-car-classification-api"
version = "0.1.0"
description = "REST API for CNN Car Classification Model"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]
